pyproject.toml
[project]
name = "voicelead-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "deepgram-sdk>=6.1.1",
    "elevenlabs>=2.43.0",
    "fastapi[standard]>=0.135.3",
    "groq>=1.1.2",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.26",
    "supabase>=2.28.3",
    "websockets<16",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]
